local_requirements_minimal.txt
# Minimal requirements that work with Python 3.8+
streamlit
google-generativeai
openai
numpy
PyPDF2
python-docx